% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-inflation.R
\name{adjustDollarForInflation}
\alias{adjustDollarForInflation}
\title{adjustDollarForInflation}
\usage{
adjustDollarForInflation(mydollar, myyear, newyear, idf = inflation.df)
}
\arguments{
\item{mydollar}{current dollar value}

\item{myyear}{current year value (4-digit format)}

\item{newyear}{new year value (4-digit format)}

\item{idf}{inflation data frame}
}
\value{
dollar, updated (adjusted from myyear to newyear)
}
\description{
adjustDollarForInflation
}
\examples{
loadInflationData();
adjustDollarForInflation( 123, 1943, 2000 ); # $123 in 1943 is about $1224.31 in 2000
adjustDollarForInflation( 123, 2000, 1943 ); # $123 in 2000 is about $12.36 in 1943

}
\seealso{
Other Inflation: 
\code{\link{lookupInflationDollar}()},
\code{\link{standardizeDollarsInDataFrame}()}
}
\concept{Inflation}
