% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-maths.R
\name{buildBoundingBoxFromRadiusAndGivenLatitudeLongitude}
\alias{buildBoundingBoxFromRadiusAndGivenLatitudeLongitude}
\title{buildBoundingBoxFromRadiusAndGivenLatitudeLongitude}
\usage{
buildBoundingBoxFromRadiusAndGivenLatitudeLongitude(
  my.radius,
  my.latitude,
  my.longitude,
  my.units = "mi"
)
}
\arguments{
\item{my.radius}{numeric}

\item{my.latitude}{numeric}

\item{my.longitude}{numeric}

\item{my.units}{a valid "length" measure from measurements:conv_unit}
}
\value{
a numeric vector of length 4: c(latitude.lower, latitude.upper, longitude.lower, longitude.upper);
}
\description{
A store locator or other "element" locator can utilize a
 bounding box to generate an initial subset; from which
 final circle radius calculations can be performed.
}
\examples{
buildBoundingBoxFromRadiusAndGivenLatitudeLongitude(10, 46.76551, -117.1919, "mi");
buildBoundingBoxFromRadiusAndGivenLatitudeLongitude(10, 46.76551, -117.1919, "km");
}
