% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-stats.R
\name{calculateZscores}
\alias{calculateZscores}
\title{calculateZscores}
\usage{
calculateZscores(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
z (numeric vector)
}
\description{
Calculate z-scores using formula:  (x - x.bar) / s.hat;
}
\examples{
calculateZscores( 1:5 );
calculateZscores( 1:9 );

}
