% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{callOrderFunctionWithMatrixInput}
\alias{callOrderFunctionWithMatrixInput}
\title{callOrderFunctionWithMatrixInput}
\usage{
callOrderFunctionWithMatrixInput(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
this is placed in a dataframe to re-order it.
}
\description{
callOrderFunctionWithMatrixInput
}
\examples{
See https://stackoverflow.com/questions/63801018/
}
