% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{convertDateStringToFormat}
\alias{convertDateStringToFormat}
\title{convertDateStringToFormat}
\usage{
convertDateStringToFormat(
  strvec,
  format.out = "\%Y",
  format.in = "\%Y-\%m-\%d \%H:\%M:\%S",
  numeric = TRUE
)
}
\arguments{
\item{strvec}{one or more strings, such as "3/24/2010 18:33"}

\item{format.out}{how you want to return the date "%Y" is just Year}

\item{format.in}{how the input is formatted, default is %Y-%m-%d %H:%M:%S}

\item{numeric}{if TRUE (default), will return numeric form.}
}
\value{
vector of numbers of same length as strvec, so can be easily 
used in dataframe environment
}
\description{
basically wraps strptime and strftime into this single call
}
\details{
I have this vector of dates in string format;
I want to convert it to this format (numeric), 
and currently, they are of this format.
}
\examples{
date.strings = c("3/24/2010 18:33", "9/3/2009 17:28", "10/14/2009 11:40", 
"7/3/2015 11:16","11/18/2010 1:29","4/23/2011 0:08","10/6/2010 11:13",
"7/26/2009 13:23","4/9/2008 13:40","8/20/2008 11:32");
years = convertDateStringToFormat(date.strings,"\%Y","\%m/\%d/\%Y \%H:\%M");
weeks = convertDateStringToFormat(date.strings,"\%W","\%m/\%d/\%Y \%H:\%M");
days = convertDateStringToFormat(date.strings,"\%j","\%m/\%d/\%Y \%H:\%M");

Ymd = convertDateStringToFormat(date.strings,"\%Y-\%m-\%d","\%m/\%d/\%Y \%H:\%M",numeric=F);

}
