% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-latlong.R
\name{convertDMStoDEC}
\alias{convertDMStoDEC}
\title{convertDMStoDEC}
\usage{
convertDMStoDEC(
  degrees,
  minutes = 0,
  seconds = 0,
  direction = "N",
  format = NULL
)
}
\arguments{
\item{degrees}{numeric (integer likely)}

\item{minutes}{numeric (integer likely)}

\item{seconds}{numeric (integer ??)}

\item{direction}{one of : c("N","S","E","W");}

\item{format}{if degrees contains a character form of the DMS, and format is
not NULL, a parser will take the character form, and build it out correctly}
}
\value{
updated form as numeric (decimal)
}
\description{
convertDMStoDEC
}
\examples{
convertDMStoDEC(30,60,45,"S");
convertDMStoDEC(-30,60,45,"S");  # negative is only working correctly on degrees

convertDMStoDEC(30,60,45,"N");
convertDMStoDEC(-30,60,45,"N");

convertDMStoDEC("105",direction="E");
convertDMStoDEC("105",direction="W");


convertDMStoDEC("105 22 48",direction="E",format="measurements");
convertDMStoDEC("105 22 48",direction="W",format="MeAsUr");

convertDMStoDEC("105 22 48 E", format="MeAsUr");
convertDMStoDEC("-105 22 48 E", format="MeAsUr");
convertDMStoDEC("105 22 48 W", format="MeAsUr");
convertDMStoDEC("-105 22 48 W", format="MeAsUr");


convertDMStoDEC(30,60,45,"N");

convertDMStoDEC( convertDECtoDMS(3.1415926535897932384626, "lat", "meas"), format="meas");
}
\seealso{
Other LatLong: 
\code{\link{buildBoundingBoxFromRadiusAndGivenLatitudeLongitude}()},
\code{\link{convertDECtoDMS}()},
\code{\link{parseDMSfromFormat}()}
}
\concept{LatLong}
