% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{createDir}
\alias{createDir}
\title{createDir}
\usage{
createDir(folder, verbose = TRUE)
}
\arguments{
\item{folder}{the folder to be created}

\item{verbose}{if true details will be printed}
}
\value{

}
\description{
dir.create has a recursive option, so you may want to use that
THIS version was written years ago in another C language and ported to R.
}
\examples{
createDir("R:/monte/says/hi/again/", verbose=TRUE);
createDir("R:/monte/says/hi/again/", verbose=TRUE);
}
