% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{createDirRecursive}
\alias{createDirRecursive}
\title{createDirRecursive}
\usage{
createDirRecursive(folder, verbose = FALSE, skip = 1)
}
\arguments{
\item{folder}{the folder to be created}

\item{verbose}{if true details will be printed}

\item{skip}{how many base levels to skip, we can't createDir("C:\")}
}
\value{

}
\description{
dir.create has a recursive option, so you may want to use that
THIS version was written years ago in another C language and ported to R.
}
\examples{
createDirRecursive("R:/monte/says/hi/", verbose=TRUE);
createDirRecursive("R:/monte/says/hi/", verbose=TRUE);
}
