% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-maths.R
\name{deg2rad}
\alias{deg2rad}
\title{deg2rad}
\usage{
deg2rad(degrees)
}
\arguments{
\item{degrees}{numeric (decimal form) of degrees}
}
\value{
numeric (decimal form) of radians
}
\description{
deg2rad
}
\examples{
deg2rad(180);
deg2rad(45);
deg2rad(90);
deg2rad(60);
deg2rad(30);
}
