% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-stats.R
\name{doMode}
\alias{doMode}
\title{doMode}
\usage{
doMode(x)
}
\arguments{
\item{x}{}
}
\value{
numeric vector that contains *all* values that are modal (could be bimodal)
}
\description{
Returns *mode* of a numeric vector x
}
\details{
*mode* is the most frequent value(s) in a set of data
}
\examples{
doMode( c(1:9) );
doMode( c(1, 1:9, 9) );
doMode( c(1, 1:9, 9, 9) );

}
