% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-stats.R
\name{doModeOpposite}
\alias{doModeOpposite}
\title{doModeOpposite}
\usage{
doModeOpposite(x)
}
\arguments{
\item{x}{}
}
\value{
numeric vector that contains *all* values that are least modal
}
\description{
Returns *!mode* of a numeric vector x
}
\details{
*!mode* is the least frequent value(s) in a set of data
}
\examples{
doModeOpposite( c(1:9) );
doModeOpposite( c(1, 1:9, 9) );
doModeOpposite( c(1, 1:9, 9, 9) );
	
}
