% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-vector.R
\name{doUnique}
\alias{doUnique}
\title{areUnique}
\usage{
doUnique(x)
}
\arguments{
\item{x}{vector (numeric or character)}
}
\value{
vector, integers of indexes of "unique" or "distinct" values
}
\description{
demonstrate unique() syntax using set notation
}
\examples{
doUnique( c(1, rep(2:8,5), 9) );
doUnique( c( rep(1,3), rep(2:8,5), rep(9,7)  ) );

library(datasets);
data(iris);
head(iris);
doUnique( as.character(iris$Species) );


doUnique( iris$Sepal.Length );
doUnique( as.numeric(unlist(iris["Sepal.Length"])) );

}
\seealso{
Other Vectors: 
\code{\link{findAllIndexesWithValueInVector}()},
\code{\link{freqMax}()},
\code{\link{freqMin}()},
\code{\link{notDuplicated}()},
\code{\link{whichMaxFreq}()},
\code{\link{whichMax}()},
\code{\link{whichMinFreq}()},
\code{\link{whichMin}()}
}
\concept{Vectors}
