% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{explodeMe}
\alias{explodeMe}
\title{explodeMe}
\usage{
explodeMe(delimiter = " ", str = "hello friend")
}
\arguments{
\item{delimiter}{character(s) to delimit the split}

\item{str}{a character string to be split}
}
\value{
a character vector
}
\description{
Similar to javascript.split and php.explode
}
\examples{
str = removeWhiteSpace("    four   scores    and  seven      years     ", "[s]");
strvec = explodeMe("[s]", str);
strvec[3];
}
