% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-stats.R
\name{findOutliersUsingZscores}
\alias{findOutliersUsingZscores}
\title{findOutliersUsingZscores}
\usage{
findOutliersUsingZscores(x, zmin = -3, zmax = 3)
}
\arguments{
\item{x}{numeric vector}

\item{zmin}{what is the lower-bound cutoff to label an outlier}

\item{zmax}{what is the upper-bound cutoff to label an outlier}
}
\value{
list of various features of the outliers
}
\description{
[assuming normality.  Is that a good assumption?]
\url{https://statisticsbyjim.com/basics/outliers/}
}
\examples{
findOutliersUsingZscores( c(-5, rep(1:3,9), 9) );
findOutliersUsingZscores( c(-5,-4, rep(1:3,9), 8,9) );
findOutliersUsingZscores( c(-5, rep(1:3,9), 9), -2, 2);
findOutliersUsingZscores( c(-5,-4, rep(1:3,9), 8,9), -2, 2 );

}
\seealso{
Other Stats: 
\code{\link{calculateZscores}()},
\code{\link{doModeOpposite}()},
\code{\link{doMode}()},
\code{\link{doSampleVariance}()},
\code{\link{doStatsSummary}()},
\code{\link{findOutliersUsingIQR}()}
}
\concept{Stats}
