% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{getIndexOfDataFrameColumns}
\alias{getIndexOfDataFrameColumns}
\title{getIndexOfDataFrameColumns}
\usage{
getIndexOfDataFrameColumns(df, mycols)
}
\arguments{
\item{df}{dataframe}

\item{mycols}{names of cols to find idx's ... string or vector of strings will work}
}
\value{
NA if not found; otherwise numeric vector of the same length as mycols, in the same order as mycols
}
\description{
getIndexOfDataFrameColumns
}
\examples{
library(datasets);
data(iris);
head(iris);
mycols = c("Sepal.Width","Petal.Width");
getIndexOfDataFrameColumns(iris,mycols);

mycols = c("Petal.Length","Sepal.Length");
getIndexOfDataFrameColumns(iris,mycols);

getIndexOfDataFrameColumns(iris,"column-does-not-exist");  # NA

}
\seealso{
Other DataFrame: 
\code{\link{getIndexOfDataFrameRows}()},
\code{\link{moveColumnsInDataFrame}()},
\code{\link{removeColumnsFromDataFrame}()},
\code{\link{removeDuplicatesFromDataFrameAllColumns}()},
\code{\link{removeDuplicatesFromDataFrameUnique}()},
\code{\link{removeDuplicatesFromDataFrame}()},
\code{\link{replaceDateStringWithDateColumns}()},
\code{\link{sortDataFrameByNumericColumns}()},
\code{\link{subsetDataFrame}()}
}
\concept{DataFrame}
