% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{getIndexOfDataFrameRows}
\alias{getIndexOfDataFrameRows}
\title{getIndexOfDataFrameRows}
\usage{
getIndexOfDataFrameRows(df, mycols, myvals)
}
\arguments{
\item{df}{dataframe}

\item{mycols}{names of cols to find idx's ... string or vector of strings will work}

\item{myvals}{values for the cols (like subsetting, but just getting row indexes)}
}
\value{
NA if not found; otherwise numeric vector of the same length as mycols, in the same order as mycols
}
\description{
getIndexOfDataFrameRows
}
\examples{
library(datasets);
data(iris);
head(iris);
mycols = c("Sepal.Width","Petal.Width");
myvals = c(3.0, 0.2);
getIndexOfDataFrameRows(iris,mycols,myvals);

myvals = c(1492, 1991);
getIndexOfDataFrameRows(iris,mycols,myvals);


getIndexOfDataFrameRows(iris,"column-does-not-exist", 123);  # Throws error

}
\seealso{
Other DataFrame: 
\code{\link{getIndexOfDataFrameColumns}()},
\code{\link{moveColumnsInDataFrame}()},
\code{\link{removeColumnsFromDataFrame}()},
\code{\link{removeDuplicatesFromDataFrameAllColumns}()},
\code{\link{removeDuplicatesFromDataFrameUnique}()},
\code{\link{removeDuplicatesFromDataFrame}()},
\code{\link{replaceDateStringWithDateColumns}()},
\code{\link{sortDataFrameByNumericColumns}()}
}
\concept{DataFrame}
