% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{getKeysFromStringWithSeparator}
\alias{getKeysFromStringWithSeparator}
\title{getKeysFromStringWithSeparator}
\usage{
getKeysFromStringWithSeparator(str, sep = ",", lower.case = TRUE)
}
\arguments{
\item{str}{What is the string?}

\item{sep}{What separates elements in string?}

\item{lower.case}{Convert input to lowercase?}
}
\value{
a character vector
}
\description{
getKeysFromStringWithSeparator
}
\examples{
str = "Actor,Producer,Writer";
keys = getKeysFromStringWithSeparator(str);
keys;
}
