% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-inflation.R
\name{grabInflationDataFrame}
\alias{grabInflationDataFrame}
\title{grabInflationDataFrame}
\usage{
grabInflationDataFrame()
}
\value{
inflation data frame (idf)
}
\description{
This has caching mechanisms in the local.path environment.
}
\details{
local.path must be defined in the global scope for this function to use it

local.data.path must be defined in the global scope for this function to use it

Every month it will lookup the values again.
}
\examples{
inflation.df = grabInflationDataFrame();
}
\seealso{
Other Inflation-Raw: 
\code{\link{grabInflationURL}()},
\code{\link{parseInflationTableHTML}()},
\code{\link{parseStoreReturnInflationData}()}
}
\concept{Inflation-Raw}
