% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-inflation.R
\name{grabInflationURL}
\alias{grabInflationURL}
\title{grabInflationURL}
\usage{
grabInflationURL()
}
\value{
inflation.url
}
\description{
grabInflationURL
}
\seealso{
Other Inflation-Raw: 
\code{\link{grabInflationDataFrame}()},
\code{\link{parseInflationTableHTML}()},
\code{\link{parseStoreReturnInflationData}()}
}
\concept{Inflation-Raw}
