% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-imdb.R
\name{IMDB.getMovieInfoFromActorSearch}
\alias{IMDB.getMovieInfoFromActorSearch}
\title{IMDB.getMovieInfoFromActorSearch}
\usage{
IMDB.getMovieInfoFromActorSearch(
  ttid,
  return.cols = NULL,
  imdb = imdb.data$all.actors.movies
)
}
\arguments{
\item{ttid}{Film identifier `ttid`}

\item{return.cols}{list of columns you want to display, by default all}

\item{imdb}{If you want to apply this function to a different dataframe, it's possible}
}
\value{
dataframe of results
}
\description{
IMDB.getMovieInfoFromActorSearch
}
