% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-imdb.R
\name{IMDB.getMoviesForPerson}
\alias{IMDB.getMoviesForPerson}
\title{IMDB.getMoviesForPerson}
\usage{
IMDB.getMoviesForPerson(nmid, return.cols = NULL, imdb = imdb.data)
}
\arguments{
\item{nmid}{Person identifier `nmid`}

\item{return.cols}{list of columns you want to display, by default all}

\item{imdb}{If you want to apply this function to a different dataframe, it's possible

dim(imdb.data$all.actors.info);     # 71332      10
dim(imdb.data$all.actors.movies);   # 282062     11
# glue table ::  nmid|ttid|rank ... rank from ActorSearchMovies
dim(imdb$all.actors.rank);          # 1842305     3}
}
\value{
dataframe of results
}
\description{
IMDB.getMoviesForPerson
}
