% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-imdb.R
\name{IMDB.getPersonInfo}
\alias{IMDB.getPersonInfo}
\title{IMDB.getPersonInfo}
\usage{
IMDB.getPersonInfo(nmid, return.cols = NULL, imdb = imdb.data$all.actors.info)
}
\arguments{
\item{nmid}{Person identifier `nmid`}

\item{return.cols}{list of columns you want to display, by default all}

\item{imdb}{If you want to apply this function to a different dataframe, it's possible

# dim(imdb.data$all.actors.info);  # 71332    10}
}
\value{
dataframe of results
}
\description{
IMDB.getPersonInfo
}
