% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-imdb.R
\name{IMDB.getUniqueCharactersForPerson}
\alias{IMDB.getUniqueCharactersForPerson}
\title{IMDB.getUniqueCharactersForPerson}
\usage{
IMDB.getUniqueCharactersForPerson(
  nmid,
  imdb = imdb.data$all.movies.actors.characters
)
}
\arguments{
\item{nmid}{Person identifier `nmid`}

\item{imdb}{If you want to apply this function to a different dataframe, it's possible}
}
\value{
dataframe of results
}
\description{
This is currently a function of `$all.movies.actors.characters`
 and *NOT* `$all.actors.info`
}
