% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-imdb.R
\name{IMDB.searchPersonName}
\alias{IMDB.searchPersonName}
\title{IMDB.searchPersonName}
\usage{
IMDB.searchPersonName(
  str,
  return.cols = NULL,
  ignore.case = TRUE,
  perl = FALSE,
  imdb = imdb.data$all.actors.info
)
}
\arguments{
\item{str}{Search string with basic wildcard `*` operator}

\item{return.cols}{list of columns you want to display, by default all}

\item{ignore.case}{Defaults to TRUE, matches "mont" and "Mont"}

\item{perl}{Defaults to FALSE, convert string query to PERL regex?}

\item{imdb}{If you want to apply this function to a different dataframe, it's possible}
}
\value{
dataframe of results
}
\description{
IMDB.searchPersonName
}
