% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{implodeMe}
\alias{implodeMe}
\title{implodeMe}
\usage{
implodeMe(delimiter = " ", strvec = c("hello", "friend"))
}
\arguments{
\item{delimiter}{character(s) to unsplit with}

\item{strvec}{a character string to be unsplit}
}
\value{
a character string
}
\description{
Similar to javascript.join and php.implode
}
\examples{
implodeMe();


str = removeWhiteSpace("    four   scores    and  seven      years     ", "[s]");
strvec = explodeMe("[s]", str);
implodeMe(",", strvec);

}
