% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-system.R
\name{isClose}
\alias{isClose}
\title{isClose}
\usage{
isClose(a, b, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{a}{number(s) ... numeric vector of length 1+}

\item{b}{number(s) ... numeric vector of length 1+}

\item{tol}{what is the tolerance level of "equalish" ... defaults to what is used in the function \code(all.equal)}
}
\value{
vector of logical (bool) with the maximum length of (a,b)
}
\description{
Due to issues regarding float-point precision, this function
assesses if two numerics are equal "almost".
}
\details{
See:  https://stackoverflow.com/questions/63787649/

The structure of \code(isClose) allows it to be used in traditional subsetting notation.
}
\examples{
options(digits=22);
n1 = 0.14999999999999999;
n2 = 0.15;
n3 = 0.15000000000000002;
N = c(n1,n2,n3);

isClose(n1,n2);
isClose(n2,n3);
isClose(n1,n3);

myTol = 0.00000000000000002;
isClose(n1,n2, myTol);
isClose(n2,n3, myTol);
isClose(n1,n3, myTol);

a = sample(N, 5, replace=T);
isClose( a, n2, myTol);
b = sample(N, 5, replace=T);
isClose( n1, b, myTol);

a = sample(N, 5, replace=T);
b = sample(N, 5, replace=T); 
isClose( a, b, myTol);

a = sample(N, 9, replace=T);
b = sample(N, 5, replace=T); 
isClose( a, b, myTol);  # prints warning, returns matrix


}
