% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-inflation.R
\name{loadInflationData}
\alias{loadInflationData}
\title{loadInflationData}
\usage{
loadInflationData()
}
\description{
loadInflationData
}
\concept{Inflation

inflation.df global gets assigned to the inflation.rds (equivalent to inflation.txt)}
