% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{loadRDS}
\alias{loadRDS}
\title{loadRDS}
\usage{
loadRDS(str)
}
\arguments{
\item{str}{string as the stem of the filename.rds}
}
\value{
dataframe if the dataset exists
}
\description{
loads an internal dataset by merely passing the string
}
\examples{
personality = loadRDS("personality-raw");
inflation = loadRDS("inflation");
# cars = loadRDS("cars");
# protein = loadRDS("protein");
}
