% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-inflation.R
\name{lookupInflationDollar}
\alias{lookupInflationDollar}
\title{lookupInflationDollar}
\usage{
lookupInflationDollar(year, idf = inflation.df)
}
\arguments{
\item{year}{numeric, 4-digit year}

\item{idf}{inflation data frame}
}
\value{
numeric dollar from idf table for that year
}
\description{
lookupInflationDollar
}
\examples{
loadInflationData();  # does lookup to create ratios
lookupInflationDollar( 1943 ); # $  865,000
lookupInflationDollar( 2000 ); # $8,610,000

}
\seealso{
Other Inflation: 
\code{\link{adjustDollarForInflation}()},
\code{\link{standardizeDollarsInDataFrame}()}
}
\concept{Inflation}
