% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{moveColumnsInDataFrame}
\alias{moveColumnsInDataFrame}
\title{moveColumnsInDataFrame}
\usage{
moveColumnsInDataFrame(ndf, mycols, where, anchor)
}
\arguments{
\item{ndf}{dataframe}

\item{mycols}{names of cols to move ... string or vector of strings will work}

\item{where}{"after" places after anchor; "before" places before anchor
@param anchor anchor to which we are moving ... either name or numeric idx}
}
\value{
dataframe, updated
}
\description{
moveColumnsInDataFrame
}
\examples{

library(datasets);
data(iris);
head(iris);
df = moveColumnsInDataFrame(iris,"Species","before","Sepal.Length");
head(df);
df = moveColumnsInDataFrame(iris,c("Sepal.Length","Sepal.Width"),"after","Species");
head(df);  # might be a bug if all are moved?

}
