% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mysql.R
\name{mysql.checkServerStatusSQL}
\alias{mysql.checkServerStatusSQL}
\title{mysql.checkServerStatusSQL}
\usage{
mysql.checkServerStatusSQL(conn, sql = "SELECT CURRENT_TIMESTAMP;")
}
\arguments{
\item{conn}{mysql.connection object (technically MariaDB)}

\item{sql}{character string of SQL query}
}
\value{
FALSE if not good, the CURRENT_TIMESTAMP (by default)
}
\description{
mysql.checkServerStatusSQL
}
