% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mysql.R
\name{mysql.dbConnect}
\alias{mysql.dbConnect}
\title{mysql.dbConnect}
\usage{
mysql.dbConnect(
  user = user,
  password = password,
  dbname = dbname,
  host = host,
  timeout.secs = 60,
  save = FALSE
)
}
\arguments{
\item{user}{character string}

\item{password}{character string}

\item{dbname}{character string}

\item{host}{character string}

\item{timeout.secs}{integer, 60 is default}

\item{save}{key to cache the db connection credential into mysql.memory}
}
\value{

}
\description{
mysql.dbConnect
}
