% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mysql.R
\name{mysql.dbConnectFromMemory}
\alias{mysql.dbConnectFromMemory}
\title{mysql.dbConnectFromMemory}
\usage{
mysql.dbConnectFromMemory(save = "")
}
\arguments{
\item{save}{single character string, key to save the db info into mysql.memory}
}
\value{
FALSE if fails, a `conn` if succeeds
}
\description{
mysql.dbConnectFromMemory
}
