% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-mysql.R
\name{mysql.fetchAllSQL}
\alias{mysql.fetchAllSQL}
\title{fetchAllSQL}
\usage{
mysql.fetchAllSQL(conn, sql, save = "")
}
\arguments{
\item{conn}{mysql.connection object (technically MariaDB)}

\item{sql}{character string of SQL query}

\item{save}{name of `conn` object that can be saved for reconnection}
}
\value{
dataframe of results; possibly FALSE or an error ... it tries to grab
 the query, and if it fails, it tries to reconnect ... if it fails twice, it fails.
}
\description{
fetchAllSQL
}
