% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-vector.R
\name{notDuplicated}
\alias{notDuplicated}
\title{notDuplicated}
\usage{
notDuplicated(x)
}
\arguments{
\item{x}{vector (numeric or character)}
}
\value{
vector, integers of indexes of "unique" or "distinct" values
}
\description{
demonstrate !duplicated() syntax using set notation
}
\examples{
notDuplicated( c(1, rep(2:8,5), 9) );
notDuplicated( c( rep(1,3), rep(2:8,5), rep(9,7)  ) );

library(datasets);
data(iris);
head(iris);
notDuplicated( as.character(iris$Species) );


notDuplicated( iris$Sepal.Length );
notDuplicated( as.numeric(unlist(iris["Sepal.Length"])) );

}
\seealso{
Other Vectors: 
\code{\link{doUnique}()},
\code{\link{findAllIndexesWithValueInVector}()},
\code{\link{freqMax}()},
\code{\link{freqMin}()},
\code{\link{whichMaxFreq}()},
\code{\link{whichMax}()},
\code{\link{whichMinFreq}()},
\code{\link{whichMin}()}
}
\concept{Vectors}
