% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{numberPadLeft}
\alias{numberPadLeft}
\title{numberPadLeft}
\usage{
numberPadLeft(n, w, c = "0")
}
\arguments{
\item{n}{The 'n'umber}

\item{w}{How long the final number is to be}

\item{c}{Fill digit, default is 0}
}
\value{
string
}
\description{
When caching pages of content, useful for organization.
 (e.g., page1.html becomes page_001.html)
}
\examples{
numberPadLeft(33,1);
numberPadLeft(33,2);
numberPadLeft(33,3);
numberPadLeft(33,4);
}
