% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-latlong.R
\name{parseDMSfromFormat}
\alias{parseDMSfromFormat}
\title{parseDMSfromFormat}
\usage{
parseDMSfromFormat(str = "3 8 29.7335529232441", format = "measurements")
}
\arguments{
\item{str}{a string form of lat/long}

\item{format}{what formatter to use to parse string}
}
\value{
a list of $degrees , $minutes , $seconds, $direction
}
\description{
parseDMSfromFormat
}
\examples{
parseDMSfromFormat();
parseDMSfromFormat("3 8 29.7335529232441", "MeAsUr");
parseDMSfromFormat("-3 8 29.7335529232441");
parseDMSfromFormat("-3 8 29.7335529232441 S");
}
\seealso{
Other LatLong: 
\code{\link{buildBoundingBoxFromRadiusAndGivenLatitudeLongitude}()},
\code{\link{convertDECtoDMS}()},
\code{\link{convertDMStoDEC}()}
}
\concept{LatLong}
