% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-inflation.R
\name{parseInflationTableHTML}
\alias{parseInflationTableHTML}
\title{parseInflationTableHTML}
\usage{
parseInflationTableHTML(infl.html)
}
\arguments{
\item{infl.html}{\code{rvest} object of HTML with inflation data}
}
\value{
inflation data frame (idf)
}
\description{
parseInflationTableHTML
}
\seealso{
Other Inflation-Raw: 
\code{\link{grabInflationDataFrame}()},
\code{\link{grabInflationURL}()},
\code{\link{parseStoreReturnInflationData}()}
}
\concept{Inflation-Raw}
