% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-inflation.R
\name{parseStoreReturnInflationData}
\alias{parseStoreReturnInflationData}
\title{parseStoreReturnInflationData}
\usage{
parseStoreReturnInflationData(raw, myfile = NULL, mytxt = NULL)
}
\arguments{
\item{raw}{HTML string harvested from URL}

\item{myfile}{recent.txt}

\item{mytxt}{YYYY-MM.txt as we update lookup every month (if running locally)}
}
\value{
inflation data frame (idf)
}
\description{
parseStoreReturnInflationData
}
\seealso{
Other Inflation-Raw: 
\code{\link{grabInflationDataFrame}()},
\code{\link{grabInflationURL}()},
\code{\link{parseInflationTableHTML}()}
}
\concept{Inflation-Raw}
