% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-tests.R
\name{performBartlettSphericityTest}
\alias{performBartlettSphericityTest}
\title{performBartlettSphericityTest}
\usage{
performBartlettSphericityTest(X, n.obs = NULL, alpha = 0.05)
}
\arguments{
\item{X}{Either the numeric dataframe or the correlation matrix of the dataframe}

\item{n.obs}{If null, I assume X is a dataframe, not a correlation matrix}

\item{alpha}{Alpha level (0.05) to report findings as $msg}
}
\value{
list with $pvalue and $msg
}
\description{
performBartlettSphericityTest
}
\examples{
library(datasets);
data(iris);
head(iris);
performBartlettSphericityTest(iris[,1:4]);
performBartlettSphericityTest(iris[,1:4], alpha=0.01);
performBartlettSphericityTest(cor(iris[,1:4]), n.obs = nrow(iris));
}
