% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-tests.R
\name{performKMOTest}
\alias{performKMOTest}
\title{performKMOTest}
\usage{
performKMOTest(X)
}
\arguments{
\item{X}{Either the numeric dataframe or the correlation matrix of the dataframe}

\item{n.obs}{If null, I assume X is a dataframe, not a correlation matrix}

\item{alpha}{Alpha level (0.05) to report findings as $msg}
}
\value{
list with $pvalue and $msg
}
\description{
performKMOTest
}
\examples{
library(datasets);
data(iris);
head(iris);
performKMOTest(iris[,1:4]);

setosa = subsetDataFrame( iris, "Species", "setosa" );
performKMOTest( setosa[,1:4] );
}
