% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-tests.R
\name{performSimpleChiSquaredTest}
\alias{performSimpleChiSquaredTest}
\title{performSimpleChiSquaredTest}
\usage{
performSimpleChiSquaredTest(chi, df, alpha = 0.05)
}
\arguments{
\item{chi}{numeric (can be vector)}

\item{df}{numeric (can be vector of matching length)}

\item{alpha}{numeric alpha-level, default is 0.05}
}
\value{
list of results
}
\description{
performSimpleChiSquaredTest
}
\examples{
performSimpleChiSquaredTest(1944.5, 36);
performSimpleChiSquaredTest(12, 22);
performSimpleChiSquaredTest(22, 12);
performSimpleChiSquaredTest(22, 12, alpha=0.01);
}
