% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{printMatrix}
\alias{printMatrix}
\title{printMatrix}
\usage{
printMatrix(mat, digits = 3, zeroIsh = TRUE, z.digits = 6)
}
\arguments{
\item{mat}{numeric matrix}

\item{digits}{numeric, digits to round}

\item{zeroIsh}{if TRUE, zeroIsh is also called}

\item{z.digits}{numeric, digits to perform zeroIsh}
}
\value{
matrix, as character string
}
\description{
printMatrix
}
\examples{
x = cbind( stats::rnorm(10,0,1), stats::rnorm(10,3,1), stats::rnorm(10,9,1) );
printMatrix( stats::cov(x), 3);
printMatrix( stats::cor(x), 3);

x = matrix( c(sin(pi), cos(pi), -sin(pi), -cos(pi)), nrow=2);
zeroIsh(x);
printMatrix( x, 3 );
printMatrix( x, 22, zeroIsh = FALSE);
printMatrix( x, 22, zeroIsh = TRUE);
printMatrix( x, 22, zeroIsh = TRUE, z.digits=16);
}
