% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{printPaste}
\alias{printPaste}
\title{printPaste}
\usage{
printPaste(..., sep = " ", collapse = NULL, recycle0 = FALSE)
}
\arguments{
\item{...}{one or more R objects, to be converted to character vectors.}

\item{sep}{a character string to separate the terms}

\item{collapse}{an optional character string to separate the results}

\item{recycle0}{TRUE / FALSE}
}
\value{
prints the result
}
\description{
merges print(paste into one function
}
