% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{printPaste0}
\alias{printPaste0}
\title{printPaste0}
\usage{
printPaste0(..., collapse = NULL, recycle0 = FALSE)
}
\arguments{
\item{...}{one or more R objects, to be converted to character vectors.}

\item{collapse}{an optional character string to separate the results}

\item{recycle0}{TRUE / FALSE}
}
\value{
prints the result
}
\description{
merges print(paste0 into one function
}
