% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{readTriangularCorrelationTable}
\alias{readTriangularCorrelationTable}
\title{readTriangularCorrelationTable}
\usage{
readTriangularCorrelationTable(file)
}
\arguments{
\item{file}{file with triangular correlation table}
}
\value{
correlation matrix
}
\description{
readTriangularCorrelationTable
}
\examples{
d.cor   = readTriangularCorrelationTable("http://md5.mshaffer.com/WSU_STATS419/DRUG_USE.txt");
d.names = c( "Cigarette", "Beer", "Wine", "Liquor", "Cocaine",
             "Tranquilizer", "Medication", "Heroin", "Marijuana",
             "Hashish", "Inhalant", "Hallucinogen", "Amphetamine");
colnames(d.cor) = d.names;
rownames(d.cor) = d.names;
d.cor;
}
