% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{removeAllColumnsBut}
\alias{removeAllColumnsBut}
\alias{removeAllColumnsExcept}
\title{removeAllColumnsBut}
\usage{
removeAllColumnsBut(df, mycols)
}
\arguments{
\item{df}{dataframe}

\item{mycols}{names of cols to keep}
}
\value{
dataframe, updated
}
\description{
Remove all columns in the data frame but (except for)
 those columns listed.
}
\details{
Useful when you have a df with lots of columns.
}
\examples{
library(datasets);
data(iris);
head(iris);
dim(iris);
ndf = removeAllColumnsBut(iris, c("Petal.Length","Petal.Width"));
head(ndf);
dim(ndf);
}
