% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{removeColumnsFromDataFrame}
\alias{removeColumnsFromDataFrame}
\title{removeColumnsFromDataFrame}
\usage{
removeColumnsFromDataFrame(df, mycols)
}
\arguments{
\item{df}{dataframe}

\item{mycols}{names of cols to remove ... string or vector of strings will work}
}
\value{
dataframe, updated
}
\description{
\code{removeColumnsFromDataFrame} removes one or more columns
from a dataframe \code{df} based on the names \code{mycols}
of said columns.
}
\examples{
library(datasets);
data(iris);
head(iris);
df = removeColumnsFromDataFrame(iris,"Species");
head(df);
}
\seealso{
Other DataFrame: 
\code{\link{getIndexOfDataFrameColumns}()},
\code{\link{getIndexOfDataFrameRows}()},
\code{\link{moveColumnsInDataFrame}()},
\code{\link{removeDuplicatesFromDataFrameAllColumns}()},
\code{\link{removeDuplicatesFromDataFrameUnique}()},
\code{\link{removeDuplicatesFromDataFrame}()},
\code{\link{replaceDateStringWithDateColumns}()},
\code{\link{sortDataFrameByNumericColumns}()},
\code{\link{subsetDataFrame}()}
}
\concept{DataFrame}
