% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{removeDuplicatesFromDataFrameAllColumns}
\alias{removeDuplicatesFromDataFrameAllColumns}
\title{removeDuplicatesFromDataFrameAllColumns}
\usage{
removeDuplicatesFromDataFrameAllColumns(df)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe, updated
}
\description{
Based on the current order of the dataframe, it will remove
duplicate values in ALL columns.
}
\examples{
library(datasets);
data(iris);
head(iris);
dim(iris);
iris[c(102,143),]; # is this a duplicate or replicate?
#if a different plant, NO (generally)
#if an error in data collection (double-submission), YES

df = removeDuplicatesFromDataFrameAllColumns(iris); # personality raw
head(df);
dim(df);
}
\seealso{
Other DataFrame: 
\code{\link{getIndexOfDataFrameColumns}()},
\code{\link{getIndexOfDataFrameRows}()},
\code{\link{moveColumnsInDataFrame}()},
\code{\link{removeColumnsFromDataFrame}()},
\code{\link{removeDuplicatesFromDataFrameUnique}()},
\code{\link{removeDuplicatesFromDataFrame}()},
\code{\link{replaceDateStringWithDateColumns}()},
\code{\link{sortDataFrameByNumericColumns}()},
\code{\link{subsetDataFrame}()}
}
\concept{DataFrame}
