% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{removeDuplicatesFromDataFrameUnique}
\alias{removeDuplicatesFromDataFrameUnique}
\title{removeDuplicatesFromDataFrameUnique}
\usage{
removeDuplicatesFromDataFrameUnique(df, mycolumn)
}
\arguments{
\item{df}{dataframe}

\item{mycolumn}{name of column to look for unique/distinct values ... string}
}
\value{
dataframe, updated
}
\description{
Based on the current (sort) order of the dataframe, it will remove
duplicate values in the column.
}
\examples{
library(datasets);
data(iris);
head(iris);
dim(iris);

df = removeDuplicatesFromDataFrameUnique(iris,"Species");
head(df);
dim(df);

df = removeDuplicatesFromDataFrameUnique(iris,"Petal.Width");
head(df);
dim(df);
}
\seealso{
Other DataFrame: 
\code{\link{getIndexOfDataFrameColumns}()},
\code{\link{getIndexOfDataFrameRows}()},
\code{\link{moveColumnsInDataFrame}()},
\code{\link{removeColumnsFromDataFrame}()},
\code{\link{removeDuplicatesFromDataFrameAllColumns}()},
\code{\link{removeDuplicatesFromDataFrame}()},
\code{\link{replaceDateStringWithDateColumns}()},
\code{\link{sortDataFrameByNumericColumns}()},
\code{\link{subsetDataFrame}()}
}
\concept{DataFrame}
