% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{removeWhiteSpace}
\alias{removeWhiteSpace}
\title{removeWhiteSpace}
\usage{
removeWhiteSpace(str, replace = " ", n = 2, pre.trim = TRUE, post.trim = TRUE)
}
\arguments{
\item{str}{character string to be adjusted}

\item{replace}{what will we replace the white space with}

\item{n}{number of spaces to find and replace}

\item{pre.trim}{if TRUE, trims the string before removing white space within}

\item{post.trim}{if TRUE, trims the string after removing white space within}
}
\value{
updated adjusted string
}
\description{
removeWhiteSpace
}
\examples{
removeWhiteSpace("    four   scores    and  seven      years     ");
removeWhiteSpace("\r\n    four   scores    and  seven      years   \t\t  ");

removeWhiteSpace("    four   scores    and  seven      years     ", "");
removeWhiteSpace("    four   scores    and  seven      years     ", "[s]");

removeWhiteSpace("The quick brown fox jumps over the lazy dog", ""); # default is 2
removeWhiteSpace("The quick brown fox jumps over the lazy dog", "", n=1);
}
