% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{replaceDateStringWithDateColumns}
\alias{replaceDateStringWithDateColumns}
\title{replaceDateStringWithDateColumns}
\usage{
replaceDateStringWithDateColumns(df, mycolumn, newcols)
}
\arguments{
\item{df}{dataframe}

\item{mycolumn}{df$mycolumn is a DateString to be replaced}

\item{newcols}{vector(s) of same length as df$mycolumn that will replace column}
}
\value{
dataframe, updated
}
\description{
replaceDateStringWithDateColumns
}
\seealso{
Other DataFrame: 
\code{\link{getIndexOfDataFrameColumns}()},
\code{\link{moveColumnsInDataFrame}()},
\code{\link{removeColumnsFromDataFrame}()},
\code{\link{removeDuplicatesFromDataFrame}()},
\code{\link{sortDataFrameByNumericColumns}()}
}
\concept{DataFrame}
