% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{replaceFactorColumnWithIndicatorVariables}
\alias{replaceFactorColumnWithIndicatorVariables}
\title{replaceFactorColumnWithIndicatorVariables}
\usage{
replaceFactorColumnWithIndicatorVariables(
  df,
  source.column,
  new.column = source.column,
  use.boolean = TRUE,
  sep = ",",
  remove.original = FALSE,
  lower.case = TRUE
)
}
\arguments{
\item{df}{Data frame}

\item{source.column}{Where the factors live, it could be exhaustive
or not the function deals with that.}

\item{new.column}{The new column name will have `.i` appended to it for each
indicator variable.}

\item{use.boolean}{If TRUE, TRUE/FALSE otherwise 1/0.}

\item{sep}{If multiple factors are in the one column, what is the separator.
Useful as this will call getKeysFromStringWithSeparator (which maybe
 should live in functions-str.R)}

\item{remove.original}{If TRUE, original column is removed}

\item{lower.case}{If TRUE, indicator keys are converted to lowercase.}
}
\value{

}
\description{
I have a column with factors or string values, and I create
 indicator variables to binary-ize those factors
}
\examples{
library(datasets);
data(iris);
head(iris);
dim(iris);

ndf = replaceFactorColumnWithIndicatorVariables(iris, "Species", "is");
head(ndf);
dim(ndf);

ndf = replaceFactorColumnWithIndicatorVariables(iris, "Species", "is", FALSE);
head(ndf);
dim(ndf);
}
