% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{rotateMatrix}
\alias{rotateMatrix}
\title{rotateMatrix}
\usage{
rotateMatrix(x, clockwise = TRUE)
}
\arguments{
\item{x}{matrix}

\item{clockwise}{direction of rotation, default is clockwise}
}
\value{
matrix, updated
}
\description{
rotateMatrix
}
\examples{
m = as.matrix(structure(c(1, 0, 4, 0, 3, 0, 2, 0, 5), .Dim = c(3L, 3L)));
rotateMatrix(m);
rotateMatrix(m,FALSE);

}
