% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-matrix.R
\name{rotateMatrix180}
\alias{rotateMatrix180}
\title{rotateMatrix180}
\usage{
rotateMatrix180(x, clockwise = TRUE)
}
\arguments{
\item{x}{matrix}

\item{clockwise}{direction of rotation, default is clockwise}
}
\value{
matrix, updated
}
\description{
rotateMatrix180
}
\examples{
m = as.matrix(structure(c(1, 0, 4, 0, 3, 0, 2, 0, 5), .Dim = c(3L, 3L)));
rotateMatrix180(m);
rotateMatrix180(m,FALSE);  # equivalent

}
