% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{roundMeToString}
\alias{roundMeToString}
\title{roundMeToString}
\usage{
roundMeToString(val, digits = 3, leadingZero = TRUE, decimal.sep = ".")
}
\arguments{
\item{val}{a numeric value}

\item{digits}{number of digits to round}

\item{leadingZero}{TRUE / FALSE ... include the leading zero, if necessary}

\item{decimal.sep}{The decimal separator, default is "."}
}
\value{
string of numeric based on parameters
}
\description{
roundMeToString
}
\examples{
roundMeToString(pi, 5);
roundMeToString(3.1415926535897932384626, 9);

roundMeToString(0.9, 3);
roundMeToString(0.9, 3, FALSE);

roundMeToString(1.9, 3);
roundMeToString(1.9, 3, FALSE);

roundMeToString(-0.9, 3);
roundMeToString(-0.9, 3, FALSE);
}
