% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{sortDataFrameByNumericColumns}
\alias{sortDataFrameByNumericColumns}
\title{sortDataFrameByNumericColumns}
\usage{
sortDataFrameByNumericColumns(df, mycols, direction = "DESC")
}
\arguments{
\item{df}{dataframe}

\item{mycols}{names of cols to sort on}

\item{direction}{direction of sort "ASC" or "DESC" ... if a vector, must be of same length as mycols}
}
\value{
dataframe, updated
}
\description{
sortDataFrameByNumericColumns
}
\examples{
library(datasets);
data(iris);
df = iris[1:10,];
df$date.strings = c("3/24/2010 18:33", "9/3/2009 17:28", "10/14/2009 11:40", 
"7/3/2015 11:16","11/18/2010 1:29","4/23/2011 0:08","10/6/2010 11:13",
"7/26/2009 13:23","4/9/2008 13:40","8/20/2008 11:32");
df$year = convertDateStringToFormat(df$date.strings,"\%Y","\%m/\%d/\%Y \%H:\%M");
df$week = convertDateStringToFormat(df$date.strings,"\%W","\%m/\%d/\%Y \%H:\%M");
df$day = convertDateStringToFormat(df$date.strings,"\%j","\%m/\%d/\%Y \%H:\%M");

df = removeColumnsFromDataFrame(df,"date.strings");
    mycols = c("year","week", "day");
sortDataFrameByNumericColumns(df,mycols,"ASC");
sortDataFrameByNumericColumns(df,mycols,"DESC");

    mydirs = c("ASC","DESC","ASC");
sortDataFrameByNumericColumns(df, mycols, mydirs );

sortDataFrameByNumericColumns(df,sample(mycols),sample(mydirs) );

}
