% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-standardize.R
\name{standardizeToFactor}
\alias{standardizeToFactor}
\title{standardizeToFactor}
\usage{
standardizeToFactor(x, factor = 1)
}
\arguments{
\item{x}{numeric vector}

\item{factor}{what we will multiply by}
}
\value{
numeric vector, updated
}
\description{
Standarize a vector by multiplying by a factor
}
\examples{
standardizeToFactor ( c(2: 10), 2 );
standardizeToFactor ( c(-2, 1: 10), 1/1000 );

}
\seealso{
Other Standardize: 
\code{\link{standardizeFromOneRangeToAnother}()},
\code{\link{standardizeToMax}()},
\code{\link{standardizeToMin}()},
\code{\link{standardizeToN}()},
\code{\link{standardizeToSum}()}
}
\concept{Standardize}
