% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-stats.R
\name{standardizeToMax}
\alias{standardizeToMax}
\title{standardizeToMax}
\usage{
standardizeToMax(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector, updated
}
\description{
Standarize a vector to its maximum
}
\examples{
standardizeToMax ( c(2: 10) );
standardizeToMax ( c(-2, 1: 10) );

}
