% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-standardize.R
\name{standardizeToMin}
\alias{standardizeToMin}
\title{standardizeToMin}
\usage{
standardizeToMin(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector, updated
}
\description{
Standarize a vector to its minimum
}
\examples{
standardizeToMin ( c(2: 10) );
standardizeToMin ( c(-2, 1: 10) );

}
\seealso{
Other Standardize: 
\code{\link{standardizeFromOneRangeToAnother}()},
\code{\link{standardizeToFactor}()},
\code{\link{standardizeToMax}()},
\code{\link{standardizeToN}()},
\code{\link{standardizeToSum}()}
}
\concept{Standardize}
