% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-stats.R
\name{standardizeToN}
\alias{standardizeToN}
\title{standardizeToN}
\usage{
standardizeToN(x, count.na = F)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector, updated
}
\description{
Standarize a vector based on it's length
}
\examples{
standardizeToN ( c(2: 10) ); 
standardizeToN ( c(-2, 1: 10) ); 
	
}
