% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{storeToFile}
\alias{storeToFile}
\title{storeToFile}
\usage{
storeToFile(str, file, encoding = "UTF-8")
}
\arguments{
\item{str}{The string to store}

\item{file}{The file to store the string (it will override).}

\item{encoding}{by default UTF-8 ... Windows is ANSI_X3.4-1986 or RStudio default is ISO-8859-1}
}
\value{

}
\description{
Store a string to a file (e.g., an HTML page downloaded).
}
