% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{strlen}
\alias{strlen}
\title{strlen}
\usage{
strlen(str)
}
\arguments{
\item{str}{the character string}
}
\value{
the numeric length of said string
}
\description{
strlen
}
\examples{
strlen("3.1415926535897932384626");
strlen( pi );
}
