% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{strPadLeft}
\alias{strPadLeft}
\alias{numberPadLeft}
\title{strPadLeft}
\usage{
strPadLeft(str, final.str.len, padding = "0")
}
\arguments{
\item{str}{The 'string' (can be a number)}

\item{final.str.len}{How long the final str is to be}

\item{padding}{Fill with, default is "0" (zero)}
}
\value{
string
}
\description{
When caching pages of content, useful for organization.
 (e.g., page1.html becomes page_001.html)
}
\examples{
strPadLeft(33,1);
strPadLeft(33,2);
strPadLeft(33,3);
strPadLeft(33,4);
}
