% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{subsetDataFrame}
\alias{subsetDataFrame}
\title{subsetDataFrame}
\usage{
subsetDataFrame(
  df,
  mycols = mycols,
  comparison = "==",
  myvals = myvals,
  verbose = FALSE
)
}
\arguments{
\item{df}{dataframe}

\item{mycols}{names of cols to find idx's ... string or vector of strings will work}

\item{comparison}{by default "==" equality comparison, LHS is df and RHS is myval}

\item{myvals}{values for the cols to subset}

\item{verbose}{if TRUE, will print messages regarding subsetting}
}
\value{
NA if not correctly specified; otherwise dataframe of subset
}
\description{
subsetDataFrame
}
\examples{
library(datasets);
data(iris);
head(iris);
mycols = c("Sepal.Width","Petal.Width");
myvals = c(3.0, 0.2);
subsetDataFrame(iris, mycols, "==", myvals);
subsetDataFrame(iris, mycols, "==", myvals, verbose=TRUE);

comparison = c(">=", "<=");
subsetDataFrame(iris, mycols, comparison=comparison, myvals, verbose=TRUE);

myvals = c(1492, 1991);
subsetDataFrame(iris, mycols, "==", myvals);  # not found, returns zero rows ...

dim( subsetDataFrame(iris,"Petal.Length", "==", 1.4) );
dim( subsetDataFrame(iris,"Petal.Length", "~=", 1.4) );

df = iris;
df$sinpi = sin(pi);
dim( subsetDataFrame(df,"sinpi", "==", 0) );
dim( subsetDataFrame(df,"sinpi", "~=", 0) );  # addresses floating point issues with isClose


subsetDataFrame(iris,"column-does-not-exist", "==", 123);  # Throws warning, returns NA

}
\seealso{
Other DataFrame: 
\code{\link{getIndexOfDataFrameColumns}()},
\code{\link{getIndexOfDataFrameRows}()},
\code{\link{moveColumnsInDataFrame}()},
\code{\link{removeColumnsFromDataFrame}()},
\code{\link{removeDuplicatesFromDataFrameAllColumns}()},
\code{\link{removeDuplicatesFromDataFrameUnique}()},
\code{\link{removeDuplicatesFromDataFrame}()},
\code{\link{replaceDateStringWithDateColumns}()},
\code{\link{sortDataFrameByNumericColumns}()}
}
\concept{DataFrame}
