% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-str.R
\name{trimMe}
\alias{trimMe}
\title{trimMe}
\usage{
trimMe(str)
}
\arguments{
\item{str}{character string to be "trimmed"}
}
\value{
updated trimmed string
}
\description{
trimMe
}
\examples{
trimMe("    four   scores    and  seven      years     ");
trimMe("\r\n    four   scores    and  seven      years   \t\t  ");
}
