% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-vector.R
\name{whichMaxFreq}
\alias{whichMaxFreq}
\title{whichMaxFreq (doMode)}
\usage{
whichMaxFreq(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector that contains _all_ values that are modal (could be bimodal)
}
\description{
Returns \code{mode} of a numeric vector x
}
\details{
\code{mode} is the most frequent value(s) in a set of data
}
\examples{
whichMaxFreq( c(1:9) );
whichMaxFreq( c(1, 1:9, 9) );
whichMaxFreq( c(1, 1:9, 9, 9) );

}
\seealso{
Other Vectors: 
\code{\link{doUnique}()},
\code{\link{findAllIndexesWithValueInVector}()},
\code{\link{freqMax}()},
\code{\link{freqMin}()},
\code{\link{notDuplicated}()},
\code{\link{whichMax}()},
\code{\link{whichMinFreq}()},
\code{\link{whichMin}()}
}
\concept{Vectors}
