% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-stats.R
\name{whichMin}
\alias{whichMin}
\title{whichMin}
\usage{
whichMin(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector that contains the indexes of *all* min elements, not just the *first*
}
\description{
behaves like which.min(x) but returns multiple indexes if required.
}
\examples{
which.min( c(23, presidents[1:30], 23) );
whichMin( c(23, presidents[1:30], 23) );
	
}
