% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-vector.R
\name{whichMinFreq}
\alias{whichMinFreq}
\title{whichMinFreq (doModeOpposite)}
\usage{
whichMinFreq(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector that contains _all_ values that are least modal
}
\description{
Returns \code{!mode} of a numeric vector x
}
\details{
\code{!mode} is the least frequent value(s) in a set of data
}
\examples{
whichMinFreq( c(1:9) );
whichMinFreq( c(1, 1:9, 9) );
whichMinFreq( c(1, 1:9, 9, 9) );

}
\seealso{
Other Vectors: 
\code{\link{findAllIndexesWithValueInVector}()},
\code{\link{freqMax}()},
\code{\link{freqMin}()},
\code{\link{whichMaxFreq}()},
\code{\link{whichMax}()},
\code{\link{whichMin}()}
}
\concept{Vectors}
