% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-search.R
\name{wildcardSearch}
\alias{wildcardSearch}
\title{wildcardSearch}
\usage{
wildcardSearch(
  str,
  col.name,
  return.cols = NULL,
  ignore.case = TRUE,
  perl = FALSE,
  df = NULL
)
}
\arguments{
\item{str}{Search string with basic wildcard `*` operator}

\item{col.name}{The column to perform search on [or a character vector]}

\item{return.cols}{list of columns you want to display, by default all}

\item{ignore.case}{Defaults to TRUE, matches "mont" and "Mont"}

\item{perl}{Defaults to FALSE, convert string query to PERL regex?}

\item{df}{Dataframe to search [column]}
}
\value{
dataframe of results, empty dataframe is possible
}
\description{
By default searches a dataframe string column based on a
 wildcard `*` operator.
}
\details{
Can also just search a single character vector.
}
