% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-date.R
\name{convertDateStringToFormat}
\alias{convertDateStringToFormat}
\title{convertDateStringToFormat}
\usage{
convertDateStringToFormat(
  strvec,
  to = "\%Y",
  to.name = "years",
  from = "\%Y-\%m-\%d \%H:\%M:\%S",
  num = TRUE
)
}
\arguments{
\item{strvec}{one or more strings, such as \code{"3/24/2010 18:33"}}

\item{to}{how you want to return the date \code{"\%Y"} is just Year}

\item{to.name}{name(s) given to the \code{to} column(s)}

\item{from}{how the input is formatted, default is \code{"\%Y-\%m-\%d \%H:\%M:\%S"}}

\item{num}{if TRUE (default), will return numeric form.}
}
\value{
dataframe same length as strvec, with one or more columns
}
\description{
basically wraps strptime and strftime into this single call
}
\details{
I have this vector of dates in string format;
I want to convert it to this format (numeric),
and currently, they are of this format.
}
\concept{Date-Time}
