% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-date.R
\name{convertDateStringToFormat}
\alias{convertDateStringToFormat}
\title{convertDateStringToFormat}
\usage{
convertDateStringToFormat(
  strvec,
  to = "\%Y",
  to.name = "years",
  from = "\%Y-\%m-\%d \%H:\%M:\%S",
  num = TRUE
)
}
\arguments{
\item{strvec}{one or more strings, such as \code{"3/24/2010 18:33"}}

\item{to}{how you want to return the date \code{"\%Y"} is just Year}

\item{to.name}{name(s) given to the \code{to} column(s)}

\item{from}{format of input, default \code{"\%Y-\%m-\%d \%H:\%M:\%S"}}

\item{num}{if TRUE (default), will return numeric form.}
}
\value{
dataframe same length as strvec, with one or more columns
}
\description{
basically wraps strptime and strftime into this single call
}
\details{
I have this vector of dates in string format;
I want to convert it to this format (numeric),
and currently, they are of this format.
}
\examples{

date.strings = c("3/24/2010 18:33", "9/3/2009 17:28", "10/14/2009 11:40",
                 "7/3/2015 11:16", "11/18/2010 1:29", "4/23/2011 0:08",
                 "10/6/2010 11:13", "7/26/2009 13:23","4/9/2008 13:40",
                 "8/20/2008 11:32");

years = convertDateStringToFormat(date.strings,
                             "\%Y", "years",
                                                           "\%m/\%d/\%Y \%H:\%M");

weeks = convertDateStringToFormat(date.strings,
                             "\%W","weeks",
                                                            "\%m/\%d/\%Y \%H:\%M");

days = convertDateStringToFormat(date.strings,"
                              \%j", "days",
                                                           "\%m/\%d/\%Y \%H:\%M");

ywd = convertDateStringToFormat( date.strings,
                             c("\%Y","\%W","\%j"), c("year","week","day"),
                                                           "\%m/\%d/\%Y \%H:\%M");


Ymd = convertDateStringToFormat(date.strings,
                              "\%Y-\%m-\%d","ISO8601",
                                                 "\%m/\%d/\%Y \%H:\%M",num=FALSE);

}
\concept{Date-Time}
