% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-stats.R
\name{doModeOpposite}
\alias{doModeOpposite}
\title{doModeOpposite}
\usage{
doModeOpposite(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector that contains _all_ values that are least modal
}
\description{
Returns \code{!mode} of a numeric vector x
}
\details{
\code{!mode} is the least frequent value(s) in a set of data
}
\examples{
doModeOpposite( c(1:9) );
doModeOpposite( c(1, 1:9, 9) );
doModeOpposite( c(1, 1:9, 9, 9) );

}
\seealso{
Other Stats: 
\code{\link{calculateZscores}()},
\code{\link{doMode}()},
\code{\link{doSampleVariance}()},
\code{\link{doStatsSummary}()},
\code{\link{findOutliersUsingIQR}()},
\code{\link{findOutliersUsingZscores}()}
}
\concept{Stats}
