% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-stats.R
\name{doStatsSummary}
\alias{doStatsSummary}
\title{doStatsSummary}
\usage{
doStatsSummary(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a list of various statistical summaries
}
\description{
doStatsSummary
}
\examples{
x.norm = doStatsSummary ( rnorm(100,0,1) );
x.unif = doStatsSummary ( runif(100,0,1) );

}
\seealso{
Other Stats: 
\code{\link{calculateZscores}()},
\code{\link{doModeOpposite}()},
\code{\link{doMode}()},
\code{\link{doSampleVariance}()},
\code{\link{findOutliersUsingIQR}()},
\code{\link{findOutliersUsingZscores}()}
}
\concept{Stats}
