% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-stats.R
\name{findOutliersUsingIQR}
\alias{findOutliersUsingIQR}
\title{findOutliersUsingIQR}
\usage{
findOutliersUsingIQR(x, innerFenceFactor = 1.5, outerFenceFactor = 3)
}
\arguments{
\item{x}{numeric vector}

\item{innerFenceFactor}{typically 1.5 * IQR}

\item{outerFenceFactor}{typically 3.0 * IQR}
}
\value{
list of various features of the outliers
}
\description{
[assuming nothing about the data]
\url{https://statisticsbyjim.com/basics/outliers/}
}
\examples{
findOutliersUsingIQR( c(-5, rep(1:3,9), 9) );
findOutliersUsingIQR( c(-5,-4, rep(1:3,9), 8,9) );
findOutliersUsingIQR( c(-5, rep(1:3,9), 9), 1, 2);  # unchanging
findOutliersUsingIQR( c(-5,-4, rep(1:3,9), 8,9), 1, 2 );  # unchanging
findOutliersUsingIQR( c(-5, rep(1:3,9), 9), 2, 4);  # unchanging
findOutliersUsingIQR( c(-5,-4, rep(1:3,9), 8,9), 2, 4 );  # unchanging

}
\seealso{
Other Stats: 
\code{\link{calculateZscores}()},
\code{\link{doModeOpposite}()},
\code{\link{doMode}()},
\code{\link{doSampleVariance}()},
\code{\link{doStatsSummary}()},
\code{\link{findOutliersUsingZscores}()}
}
\concept{Stats}
