% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{grabHTML}
\alias{grabHTML}
\title{grabHTML}
\usage{
grabHTML(htmlfile, htmlurl, verbose = TRUE, return.raw = TRUE)
}
\arguments{
\item{htmlfile}{If cached, we can return this file.  If not, we can store
the url contents to a file.  (rvest doesn't collect the original data
source to possibly do offline re-parsing.  bad data provenance.)}

\item{htmlurl}{Using RCurl, grab the raw contents off the web.}

\item{verbose}{If TRUE, will provide some verbosity}

\item{return.raw}{If TRUE, will return the raw string}
}
\value{

}
\description{
This grabs an HTML file (or any web TXT file) and caches locally.
There needs to be a global variable `local.data.path` assigned so
this function understands it is running in a local environment where
it can actually save files.  As opposed to `github.data.path` I can't
source this file in a remote repository if I can't store files there.
}
