% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-imdb.R
\name{loadDataIMDB}
\alias{loadDataIMDB}
\title{loadDataIMDB}
\usage{
loadDataIMDB()
}
\description{
This data was pulled from \url{https://www.imdb.com} in September 2020.
}
\details{
It contains several dataframes.
See \url{http://md5.mshaffer.com/WSU_STATS419/imdb.html"} for details.
}
\concept{IMDB}
