% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-maths.R
\name{rad2deg}
\alias{rad2deg}
\title{deg2rad}
\usage{
rad2deg(radians)
}
\arguments{
\item{radians}{numeric (decimal form) of radians}
}
\value{
numeric (decimal form) of degrees
}
\description{
deg2rad
}
\examples{
rad2deg( deg2rad(180) );
rad2deg( deg2rad(45) );
rad2deg( deg2rad(90) );
rad2deg( deg2rad(60) );
rad2deg( deg2rad(30) );

rad2deg( pi );
rad2deg( pi/2 );
rad2deg( pi/4 );
rad2deg( pi/3 );
rad2deg( pi/6 );
}
