% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{removeDuplicatesFromDataFrame}
\alias{removeDuplicatesFromDataFrame}
\title{removeDuplicatesFromDataFrame}
\usage{
removeDuplicatesFromDataFrame(df, mycolumn)
}
\arguments{
\item{df}{dataframe}

\item{mycolumn}{name of column to look for unique/distinct values ... string}
}
\value{
dataframe, updated
}
\description{
Based on the current order of the dataframe, it will remove
duplicate values in the column.
}
\examples{
library(datasets);
data(iris);
head(iris);
dim(iris);
df = removeDuplicatesFromDataFrame(iris,"Species");
head(df);
dim(df);
}
\seealso{
Other DataFrame: 
\code{\link{getIndexOfDataFrameColumns}()},
\code{\link{getIndexOfDataFrameRows}()},
\code{\link{moveColumnsInDataFrame}()},
\code{\link{removeColumnsFromDataFrame}()},
\code{\link{removeDuplicatesFromDataFrameAllColumns}()},
\code{\link{removeDuplicatesFromDataFrameUnique}()},
\code{\link{replaceDateStringWithDateColumns}()},
\code{\link{sortDataFrameByNumericColumns}()},
\code{\link{subsetDataFrame}()}
}
\concept{DataFrame}
