% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{replaceDateStringWithDateColumns}
\alias{replaceDateStringWithDateColumns}
\title{replaceDateStringWithDateColumns}
\usage{
replaceDateStringWithDateColumns(df, mycolumn, newcols)
}
\arguments{
\item{df}{dataframe}

\item{mycolumn}{df$mycolumn is a DateString to be replaced}

\item{newcols}{vector(s) of same length as df$mycolumn that will replace column}
}
\value{
dataframe, updated
}
\description{
replaceDateStringWithDateColumns
}
\examples{
library(datasets);
data(iris);
df = iris[1:10,];
df$date.strings = c("3/24/2010 18:33", "9/3/2009 17:28", "10/14/2009 11:40", 
"7/3/2015 11:16","11/18/2010 1:29","4/23/2011 0:08","10/6/2010 11:13",
"7/26/2009 13:23","4/9/2008 13:40","8/20/2008 11:32");
years = convertDateStringToFormat(df$date.strings,"\%Y","\%m/\%d/\%Y \%H:\%M");
weeks = convertDateStringToFormat(df$date.strings,"\%W","\%m/\%d/\%Y \%H:\%M");
days = convertDateStringToFormat(df$date.strings,"\%j","\%m/\%d/\%Y \%H:\%M");

newcols = cbind(years,weeks,days); 
      colnames(newcols) = c("year","week","day");

replaceDateStringWithDateColumns(df,"date.strings",newcols);

}
