% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-dataframe.R
\name{sortDataFrameByNumericColumns}
\alias{sortDataFrameByNumericColumns}
\title{sortDataFrameByNumericColumns}
\usage{
sortDataFrameByNumericColumns(df, mycols, direction = "DESC")
}
\arguments{
\item{df}{dataframe}

\item{mycols}{names of cols to sort on}

\item{direction}{direction of sort "ASC" or "DESC" ... if a vector, must be of same length as mycols}
}
\value{
dataframe, updated
}
\description{
Does it have to be numeric?
}
\examples{

library(datasets);
data(iris);
df = iris[1:10,];
sortDataFrameByNumericColumns(df,"Petal.Length","ASC");
sortDataFrameByNumericColumns(df,"Petal.Length","DESC");
sortDataFrameByNumericColumns(df, c("Petal.Length","Petal.Width") , "DESC");
sortDataFrameByNumericColumns(df, c("Petal.Length","Petal.Width") , c("ASC","DESC"));


date.strings = c("3/24/2010 18:33", "9/3/2009 17:28", "10/14/2009 11:40",
                 "7/3/2015 11:16", "11/18/2010 1:29", "4/23/2011 0:08",
                 "10/6/2010 11:13", "7/26/2009 13:23","4/9/2008 13:40",
                 "8/20/2008 11:32");

df$date.strings = date.strings;
df;
df = moveColumnsInDataFrame(df, "date.strings", "after", "Sepal.Width");
df;



ywd = convertDateStringToFormat( date.strings,
                             c("\%Y","\%W","\%j"), c("year","week","day"),
                                                           "\%m/\%d/\%Y \%H:\%M");

udf = replaceDateStringWithDateColumns(df,"date.strings",ywd);

    mycols = c("year","week", "day");
sortDataFrameByNumericColumns(udf,mycols,"ASC");
sortDataFrameByNumericColumns(udf,mycols,"DESC");

    mydirections = c("ASC","DESC","ASC");
sortDataFrameByNumericColumns(udf, mycols, mydirections );

sortDataFrameByNumericColumns(udf,sample(mycols),sample(mydirections) );

}
\seealso{
Other DataFrame: 
\code{\link{getIndexOfDataFrameColumns}()},
\code{\link{getIndexOfDataFrameRows}()},
\code{\link{moveColumnsInDataFrame}()},
\code{\link{removeColumnsFromDataFrame}()},
\code{\link{removeDuplicatesFromDataFrameAllColumns}()},
\code{\link{removeDuplicatesFromDataFrameUnique}()},
\code{\link{removeDuplicatesFromDataFrame}()},
\code{\link{replaceDateStringWithDateColumns}()}
}
\concept{DataFrame}
