% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-standardize.R
\name{standardizeFromOneRangeToAnother}
\alias{standardizeFromOneRangeToAnother}
\title{standardizeFromOneRangeToAnother}
\usage{
standardizeFromOneRangeToAnother(x, another, onerange = range(x))
}
\arguments{
\item{x}{numeric vector}

\item{another}{this is the new scaled range [numeric vector of length 2]}

\item{onerange}{this is typically the current range of the vector [numeric vector of length 2]}
}
\value{
numeric vector, updated
}
\description{
\url{https://stats.stackexchange.com/questions/281162/scale-a-number-between-a-range}
}
\examples{
standardizeFromOneRangeToAnother( 1:5, c(0,1) );  # map to [0,1]
standardizeFromOneRangeToAnother( runif(20, min=-1, max=1), c(0,1) );  # map to [0,1]
standardizeFromOneRangeToAnother( 1:5, c(1,7) );  # e.g., Likert-5 to Likert-7
standardizeFromOneRangeToAnother( rep(1,10), c(0,1) );  #NaN ... division by zero

}
\seealso{
Other Standardize: 
\code{\link{standardizeToFactor}()},
\code{\link{standardizeToMax}()},
\code{\link{standardizeToMin}()},
\code{\link{standardizeToN}()},
\code{\link{standardizeToSum}()}
}
\concept{Standardize}
