% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-standardize.R
\name{standardizeToMax}
\alias{standardizeToMax}
\title{standardizeToMax}
\usage{
standardizeToMax(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector, updated
}
\description{
Standarize a vector to its maximum
}
\examples{
standardizeToMax ( c(2: 10) );
standardizeToMax ( c(-2, 1: 10) );

}
\seealso{
Other Standardize: 
\code{\link{standardizeFromOneRangeToAnother}()},
\code{\link{standardizeToFactor}()},
\code{\link{standardizeToMin}()},
\code{\link{standardizeToN}()},
\code{\link{standardizeToSum}()}
}
\concept{Standardize}
