% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-standardize.R
\name{standardizeToN}
\alias{standardizeToN}
\title{standardizeToN}
\usage{
standardizeToN(x, count.na = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{count.na}{should we use NA's in our count for normalization?}
}
\value{
numeric vector, updated
}
\description{
Standarize a vector based on it's length
}
\examples{
standardizeToN ( c(2: 10) );
standardizeToN ( c(-2, 1: 10) );

}
\seealso{
Other Standardize: 
\code{\link{standardizeFromOneRangeToAnother}()},
\code{\link{standardizeToFactor}()},
\code{\link{standardizeToMax}()},
\code{\link{standardizeToMin}()},
\code{\link{standardizeToSum}()}
}
\concept{Standardize}
