% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-standardize.R
\name{standardizeToSum}
\alias{standardizeToSum}
\title{standardizeToSum}
\usage{
standardizeToSum(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector, updated
}
\description{
Standarize a vector based on it's sum
}
\examples{
s.s = standardizeToSum ( c(2: 10) );    s.s;
sum(s.s);

s.s = standardizeToN ( c(-2, 1: 10) );  s.s;
sum(s.s);

}
\seealso{
Other Standardize: 
\code{\link{standardizeFromOneRangeToAnother}()},
\code{\link{standardizeToFactor}()},
\code{\link{standardizeToMax}()},
\code{\link{standardizeToMin}()},
\code{\link{standardizeToN}()}
}
\concept{Standardize}
