% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-file.R
\name{storeToFile}
\alias{storeToFile}
\title{storeToFile}
\usage{
storeToFile(str, myfile)
}
\arguments{
\item{str}{The string to store}

\item{myfile}{The file to store the string (it will override).}
}
\value{

}
\description{
Store a string to a file (e.g., an HTML page downloaded).
}
